// src/types.ts
export enum OccasionType {
  BIRTHDAY = "День Рождения",
  NEW_YEAR = "Новый Год",
}

export enum ToneType {
  OFFICIAL = "Официальный",
  FRIENDLY = "Дружеский",
  HUMOROUS = "Юмористический",
  ROMANTIC = "Романтический",
  TOUCHING = "Трогательный",
  ADULT = "18+",
}

export type LanguageType = "Русский" | "Белорусский" | "English" | "Deutsch" | "Español" | "Français";